\name{search.massbank_eu}
\alias{search.massbank_eu}
\title{Deceleration of massbank database by compound name or other informations.(●'◡'●)}
\usage{
search.massbank_eu()
}
\description{

Tell yourself once a day, "I'm really good."

}
\examples{
search.massbank_eu(compound_name = "Dihydrotestosterone",
                   formula = NULL,
                   Instrument_Type = "LC-ESI-QTOF",
                   ms2_matrix = ms2_matrix_demo())
}
